[bugfix RELEASE] fix assertion when polymorphic relatoinship key absent
This fixes an issue where missing polymorphic keys in a payload would
trigger an assertion.

For example, if a Book model had a `hasMany` relationship to the Author
model, but the `authors` key was missing in the payload for a
polymorphic relationship, an assertion was incorrectly triggered.

This worked before 1.13, and this commit restores that functionality.

Fixes emberjs/data#3630