[bugfix RELEASE] Handle empty polymorphic relationships in payload
The polymorphic relationship was `null` when it was sent back from the
server. For example, given a payload like:

```javascript
{
  "doomsday_devices": [
    {
      "id": "1",
      "evil_minion": null
    }
  ]
}
```

Ember Data would try to extract the relationship information for
EvilMinion, but errored out since EvilMinion was `null`. The previous
check for null was insufficient because in JavaScript, typeof null
=== 'object'.

The check was switched to explicitly check for null before checking
for it to be an object.