Resolve "`this.validateFullName` is undefined" Error
When upgrading an app to Ember 2.0.2 and Ember Data I noticed the
following error:

    Uncaught TypeError: this.validateFullName is not a function

This was occurring in the registry here:

    https://github.com/emberjs/ember.js/blob/58093a5521c3a0b58fa90dae6afaafb79a9688ab/packages/container/lib/registry.js#L162

This error stems from the fact that the registration occurs in the wrong
context. Instead of a Registry, it's an Application.

The more I looked at this instance initializer though, the more I
couldn't figure out why it was here even to begin with as the
initializer has already registered the serializer and adapter. A
reasonable solution is, therefore, to flat out nuke the offending code.