Do not publish in CI anymore
AFAICT, this code would publish to bower and npm.

Publishing to bower in not needed anymore.

Regading npm... The package was last published three years ago. Odds are
high something will get wrong while publishing in CI.

For now, let's prefer manual release instead. If we'd like to add the
feature back, we could do it with a consolidated solution like
release-it.