Remove ModelRegistry Insanity
I really don't understand why this is here. We want a String, we know it
should be a String. Leaving the original code in place resulted in weird
`never` types being inferred. I made the decision to use this signature
after examining the upstream types for Ember Data's JSON:API adapter and
noting this simplified signature.